archive/v1/requirements_v1.txt
requests==2.32.3
pandas==2.2.3
duckdb==1.1.3
openai==1.57.2
typer==0.15.1
rich==13.9.4
prompt-toolkit==3.0.48
python-dotenv==1.0.1
polars==0.20.3
pyyaml==6.0.1
flask==3.0.0
gunicorn==21.2.0
scipy==1.11.4
numpy
python-telegram-bot==21.7
python-dateutil==2.8.2
aiohttp==3.9.5
web3==6.13.0
solana==0.30.2
base58==2.1.1
Flask==3.0.0
colorama==0.4.6
tabulate==0.9.0
click==8.1.7
solders==0.18.1
asyncio==3.4.3
httpx==0.26.0
matplotlib==3.8.2
seaborn==0.13.1 